feat: 테스트 디렉토리 구조 변경 및 BoardGroupServiceTest 리팩터링
- 테스트 메서드를 명확하게 분리할 수 있도록 @Nested 어노테이션 적용
- PrameterizedTest에서 인자를 재사용할수 있도록 @ValueSource 대신 @MethodSource 적용
